+++ packages/ciphernode/e3_request/src/lib.rs
@@ -1,11 +1,11 @@
-mod committee_meta;
 mod context;
 mod hetrogenous_map;
+mod meta;
 mod repo;
 mod router;
 
-pub use committee_meta::*;
 pub use context::*;
 pub use hetrogenous_map::*;
+pub use meta::*;
 pub use repo::*;
 pub use router::*;
